Move protein info to subdirectory
